Fix/post service

+++ src/main/java/org/example/service/PostLikeService.java
@@ -1,7 +1,11 @@
 package org.example.service;
 
 import lombok.AllArgsConstructor;
+import org.example.exception.exceptions.PostNotFoundException;
+import org.example.model.entity.Post;
 import org.example.properties.TopicProperties;
+import org.example.utils.DatabaseContextHolder;
+import org.example.utils.DatabaseType;
 import org.springframework.beans.factory.annotation.Value;
 import org.example.events.LikeNotificationEvent;
 import org.example.model.entity.PostLike;
@@ -16,19 +20,42 @@ public class PostLikeService {
     private final PostRepository postRepository;
     private final KafkaTemplate<String, LikeNotificationEvent> kafkaTemplate;
     private final TopicProperties topicProperties;
-    public void likePost(Long postId, Long userId) {
-        if (!likeRepository.existsByPostIdAndUserId(postId, userId)) {
-            likeRepository.save(new PostLike(null,userId ,postRepository.getById(postId)));
-            LikeNotificationEvent event = new LikeNotificationEvent(postRepository.findById(postId).get().getUserId(), postId, userId);
-            kafkaTemplate.send(topicProperties.getLikePost(), event);
+    public boolean toggleLike(Long postId, Long userId) {
+        DatabaseContextHolder.setDatabaseType(DatabaseType.MASTER);
+        try {
+            boolean alreadyLiked = likeRepository.existsByPostIdAndUserId(postId, userId);
+            if (alreadyLiked) {
+                likeRepository.deleteByPostIdAndUserId(postId, userId);
+                return false;
+            } else {
+                Post post = postRepository.findById(postId)
+                        .orElseThrow(() -> new PostNotFoundException("Post not found"));
+                likeRepository.save(new PostLike(userId, post));
+
+                LikeNotificationEvent event = new LikeNotificationEvent(post.getUserId(), postId, userId);
+                kafkaTemplate.send(topicProperties.getLikePost(), event);
+                return true;
+            }
+        } finally {
+            DatabaseContextHolder.clear();
         }
     }
 
-    public void unlikePost(Long postId, Long userId) {
-        likeRepository.deleteByPostIdAndUserId(postId, userId);
+    public boolean isPostLikedByUser(Long postId, Long userId) {
+        DatabaseContextHolder.setDatabaseType(DatabaseType.SLAVE);
+        try {
+            return likeRepository.existsByPostIdAndUserId(postId, userId);
+        } finally {
+            DatabaseContextHolder.clear();
+        }
     }
 
     public long getLikesCount(Long postId) {
-        return likeRepository.countByPostId(postId);
+        DatabaseContextHolder.setDatabaseType(DatabaseType.SLAVE);
+        try {
+            return likeRepository.countByPostId(postId);
+        }finally {
+            DatabaseContextHolder.clear();
+        }
     }
 }
